Refactored CSS structure to make it scalable for more views and themes
