+++ src/App.js
@@ -1,11 +1,11 @@
 import React, { Component } from 'react';
-import './css/App.scss';
 import Card from './components/Card';
+import './css/components/App.scss';
 
 class App extends Component {
   render() {
     return (
-      <Card name='company name' dayLimit='90' theme='minimal' />
+      <Card name='company name' dayLimit='90' theme='' />
     );
   }
 }
