--- src/css/App.scss
@@ -1,9 +1,0 @@
-*, html { padding: 0; margin: 0; color: #f5f5f5; }
-body {
-  background: #14131A;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-  font-size: 15px;
-  padding-top: 20vh;
-  display: flex;
-  justify-content: center;
-}
