+++ src/css/base.scss
@@ -1,0 +1,21 @@
+// Basic reset
+
+* { padding: 0; margin: 0; -moz-box-sizing: border-box; box-sizing: border-box; }
+
+html{ font-size: 62.5% }
+
+body {
+  margin: 0;
+  padding: 0;
+  font: $base-font-weight $base-font-size $base-font-family;
+  font-kerning: normal;
+  -webkit-text-size-adjust: 100%;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  -webkit-font-feature-settings: "kern" 1;
+     -moz-font-feature-settings: "kern" 1;
+       -o-font-feature-settings: "kern" 1;
+          font-feature-settings: "kern" 1;
+  scroll-behavior: smooth;
+}
+code { font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New", monospace; }
