+++ src/css/components/App.scss
@@ -1,0 +1,9 @@
+// User account front-page
+
+body {
+  padding-top: 20vh;
+  display: flex;
+  justify-content: center;
+}
+
+@import "../themes/default"
