+++ src/css/components/Card.scss
@@ -1,8 +1,9 @@
 .card {
   margin: 0 auto;
   padding: 16px 20px 20px;
-  background: #1D1C2D;
+  background: rgba(31, 48, 65, 0.75);
   border-radius: 3px;
+  border: 1px solid #21729C;
 }
 .card-title {
   margin: 0 0 16px;
