+++ src/css/themes/default.scss
@@ -1,0 +1,11 @@
+// Variables
+
+$theme-font-color: #FFF;
+$theme-background: #1E2129;
+
+// Default theme
+
+body {
+  color: $theme-font-color;
+  background: $theme-background;
+}
